# openapi.yaml
openapi: '3.0.3'
info:
  title: Fotografija API
  description: REST API for managing photoshoots, photos, and comments
  version: '1.0.0'
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://129.151.213.227:3000/api
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token from NextAuth
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        email:
          type: string
          format: email
        username:
          type: string
        role:
          type: string
          enum: [USER, ADMIN]
    Photoshoot:
      type: object
      required:
        - id
        - title
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          maxLength: 50
          example: "Summer Wedding 2024"
        description:
          type: string
          nullable: true
          example: "Beautiful outdoor wedding ceremony"
        date:
          type: string
          format: date-time
          nullable: true
          example: "2024-06-15T14:30:00Z"
        ownerId:
          type: integer
          nullable: true
          example: 1
        public:
          type: boolean
          default: false
          description: Whether the photoshoot is publicly visible
        sharedWith:
          type: array
          items:
            type: integer
          description: List of user IDs with whom the photoshoot is shared
        created_at:
          type: string
          format: date-time
          example: "2024-10-06T10:30:00Z"
        updated_at:
          type: string
          format: date-time
          example: "2024-10-06T10:30:00Z"
    CreatePhotoshoot:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          maxLength: 50
          example: "Summer Wedding 2024"
        description:
          type: string
          nullable: true
          example: "Beautiful outdoor wedding ceremony"
        date:
          type: string
          format: date-time
          nullable: true
          example: "2024-06-15T14:30:00Z"
        public:
          type: boolean
          default: false
        sharedWith:
          type: array
          items:
            type: integer
          description: List of user IDs to share with
    UpdatePhotoshoot:
      type: object
      properties:
        title:
          type: string
          maxLength: 50
          example: "Summer Wedding 2024"
        description:
          type: string
          nullable: true
          example: "Beautiful outdoor wedding ceremony"
        date:
          type: string
          format: date-time
          nullable: true
          example: "2024-06-15T14:30:00Z"
        ownerId:
          type: integer
          nullable: true
          example: 1
...
    Comment:
      type: object
      required:
        - id
        - photoId
        - authorId
        - body
        - created_at
      properties:
        id:
          type: integer
          example: 1
        photoId:
          type: integer
          example: 1
        authorId:
          type: integer
          example: 1
        body:
          type: string
          example: "Beautiful composition!"
        created_at:
          type: string
          format: date-time
          example: "2024-10-06T10:30:00Z"
    CreateComment:
      type: object
      required:
        - authorId
        - body
      properties:
        authorId:
          type: integer
          example: 1
        body:
          type: string
          example: "Beautiful composition!"
...
paths:
...
  /photoshoots:
    get:
      summary: Get all photoshoots
      description: Returns all photoshoots. Supports filtering by public status and owner ID.
      operationId: listPhotoshoots
      tags:
        - Photoshoots
      parameters:
        - name: public
          in: query
          description: Filter by public visibility
          schema:
            type: boolean
        - name: ownerId
          in: query
          description: Filter by owner ID
          schema:
            type: integer
      responses:
        '200':
          description: List of photoshoots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Photoshoot'
        '500':
          description: Server error
          
    post:
      summary: Create a new photoshoot
      description: Creates a new photoshoot. Requires authentication.
      operationId: createPhotoshoot
      tags:
        - Photoshoots
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePhotoshoot'
      responses:
        '201':
          description: Photoshoot created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photoshoot'
        '401':
          description: Authentication required
        '409':
          description: Photoshoot with this title already exists
        '422':
          description: Validation error
        '500':
          description: Server error
  /photoshoots/{photoshootId}:
    parameters:
      - name: photoshootId
        in: path
        required: true
        description: Photoshoot ID
        schema:
          type: integer
    
    get:
      summary: Get a specific photoshoot
      description: Returns details of a specific photoshoot
      operationId: getPhotoshoot
      tags:
        - Photoshoots
      responses:
        '200':
          description: Photoshoot details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photoshoot'
        '400':
          description: Invalid photoshoot ID
        '404':
          description: Photoshoot not found
        '500':
          description: Server error
          
    put:
      summary: Update a photoshoot
      description: Updates an existing photoshoot. Owner or admin can update.
      operationId: updatePhotoshoot
      tags:
        - Photoshoots
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePhotoshoot'
      responses:
        '200':
          description: Photoshoot updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Photoshoot'
        '400':
          description: Invalid photoshoot ID
        '401':
          description: Authentication required
        '403':
          description: Forbidden - Must be owner or admin
        '404':
          description: Photoshoot not found
        '422':
          description: Validation error
        '500':
          description: Server error
          
    delete:
      summary: Delete a photoshoot
      description: Deletes a photoshoot. Owner or admin can delete.
      operationId: deletePhotoshoot
      tags:
        - Photoshoots
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Photoshoot deleted
        '400':
          description: Invalid photoshoot ID
        '401':
          description: Authentication required
        '403':
          description: Forbidden - Must be owner or admin
        '404':
          description: Photoshoot not found
        '500':
          description: Server error
...
  /photoshoots/{photoshootId}/photos/{photoId}/comments:
    parameters:
      - name: photoshootId
        in: path
        required: true
        description: Photoshoot ID
        schema:
          type: integer
      - name: photoId
        in: path
        required: true
        description: Photo ID
        schema:
          type: integer
    
    get:
      summary: Get all comments for a photo
      description: Returns all comments on a specific photo
      operationId: listComments
      tags:
        - Comments
      responses:
        '200':
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid ID
        '404':
          description: Photo or photoshoot not found
        '500':
          description: Server error
          
    post:
      summary: Create a new comment on a photo
      description: Adds a comment to a photo. Requires authentication.
      operationId: createComment
      tags:
        - Comments
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateComment'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          description: Invalid ID
        '404':
          description: Photo or photoshoot not found
        '422':
          description: Validation error
        '500':
          description: Server error
...
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        description: User ID
        schema:
          type: integer
    
    get:
      summary: Get user by ID
      description: Returns basic user information
      operationId: getUser
      tags:
        - Users
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
    
    put:
      summary: Update user
      description: Updates user information. Requires authentication.
      operationId: updateUser
      tags:
        - Users
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                phoneNumber:
                  type: string
      responses:
        '200':
          description: User updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          description: Authentication required
        '403':
          description: Forbidden
        '404':
          description: User not found
        '422':
          description: Validation error
    
    delete:
      summary: Delete user
      description: Deletes a user. Admin only.
      operationId: deleteUser
      tags:
        - Users
      security:
        - BearerAuth: []
      responses:
        '204':
          description: User deleted
        '401':
          description: Authentication required
        '403':
          description: Forbidden - Admin role required
        '404':
          description: User not found